+++ WalletWasabi.Tests/UnitTests/UserInterfaceTest/PocketSelectionTests.cs
@@ -2,7 +2,6 @@ using System.Collections.Generic;
 using System.Linq;
 using NBitcoin;
 using WalletWasabi.Blockchain.Analysis.Clustering;
-using WalletWasabi.Blockchain.Keys;
 using WalletWasabi.Blockchain.TransactionOutputs;
 using WalletWasabi.Fluent.Helpers;
 using WalletWasabi.Fluent.Models;
@@ -1030,53 +1029,5 @@ public class PocketSelectionTests
 		recipient = "Doesn't matter, it will use semi private coins";
 		output = selection.AutoSelectPockets(recipient);
 		Assert.False(selection.IsOtherSelectionPossible(output.SelectMany(x => x.Coins), recipient));
-	}
-}
-
-internal static class LabelTestExtensions
-{
-	private static readonly KeyManager KeyManager = ServiceFactory.CreateKeyManager();
-
-	public static LabelViewModel GetLabel(this LabelSelectionViewModel selection, string label)
-	{
-		return selection.AllLabelsViewModel.Single(x => x.Value == label);
-	}
-
-	public static void AddPocket(this List<Pocket> pockets, decimal amount, out Pocket pocket, params string[] labels)
-	{
-		var label = new SmartLabel(labels);
-		var coinsView = new CoinsView(new[] { BitcoinFactory.CreateSmartCoin(NewKey(label), amount) });
-		pocket = new Pocket((label, coinsView));
-		pockets.Add(pocket);
-	}
-
-	public static HdPubKey NewKey(string label = "", int anonymitySet = 1)
-	{
-		var key = KeyManager.GenerateNewKey(label, KeyState.Used, true, false);
-		key.SetAnonymitySet(anonymitySet);
-		key.SetLabel(label);
-
-		return key;
-	}
-
-	public static SmartCoin CreateCoin(decimal amount, string label = "", int anonymitySet = 1)
-	{
-		var coin = BitcoinFactory.CreateSmartCoin(NewKey(label: label, anonymitySet: anonymitySet), amount);
-		coin.HdPubKey.SetAnonymitySet(anonymitySet);
-
-		return coin;
-	}
-
-	public static Pocket CreateSingleCoinPocket(decimal amount, string label = "", int anonSet = 0)
-	{
-		var coins = new[]
-		{
-			CreateCoin(amount, label, anonSet)
-		};
-
-		var coinsView = new CoinsView(coins.ToArray());
-		var pocket = new Pocket((label, coinsView));
-
-		return pocket;
 	}
 }
